Removed warnings when crate is compiled without flags (#847)
