+++ src/compute/mod.rs
@@ -57,9 +57,6 @@ pub mod like;
 #[cfg(feature = "compute_limit")]
 #[cfg_attr(docsrs, doc(cfg(feature = "compute_limit")))]
 pub mod limit;
-#[cfg(feature = "compute_lower")]
-#[cfg_attr(docsrs, doc(cfg(feature = "compute_lower")))]
-pub mod lower;
 #[cfg(feature = "compute_merge_sort")]
 #[cfg_attr(docsrs, doc(cfg(feature = "compute_merge_sort")))]
 pub mod merge_sort;
@@ -84,9 +81,9 @@ pub mod take;
 #[cfg(feature = "compute_temporal")]
 #[cfg_attr(docsrs, doc(cfg(feature = "compute_temporal")))]
 pub mod temporal;
-#[cfg(feature = "compute_upper")]
-#[cfg_attr(docsrs, doc(cfg(feature = "compute_upper")))]
-pub mod upper;
+#[cfg(feature = "compute_utf8")]
+#[cfg_attr(docsrs, doc(cfg(feature = "compute_utf8")))]
+pub mod utf8;
 mod utils;
 #[cfg(feature = "compute_window")]
 #[cfg_attr(docsrs, doc(cfg(feature = "compute_window")))]
